Update require path cause of rename folder
